All major updates to the script editing system
The script system is now fully operational but still requires minor bug fixes and tweaks. The next big update will consist of introducing the Visual part of the former panel into the Panel tab. Last will be a result viewer with 'hopefully' a file tree system.

+++ miezepy/core/mask_modules/mask_shape.py
@@ -84,7 +84,7 @@ class MaskShape:
         
         temp_draw.polygon(tuple([tuple(edge) for edge in edges]), fill = 1)
 
-        return np.array(temp_image)
+        return np.array(temp_image).astype(np.int16)
 
     def processSector(self, radius_range, angle_range, size_x, size_y):
         '''
